feat(ir): add spill/reload pseudo-ops
These opcodes do not correspond to real instructions; they are used by a
special pair of passes that use these ops to communicate information
about SSA values which need to be spilled to memory temporarily, in
order to split the live ranges of those values.

The first pass places the spill and reload ops where they are needed,
using the results of an analysis that determines that placement based on
the operand stack pressure of the original IR.

The IR produced by the first pass is still technically valid, but is only
partially transformed - the new SSA values produced by the reload ops are
unused, so the live ranges we aimed to split by introducing the spills and
reloads are not yet split. That brings us to the second pass.

The second pass rewires all uses of spilled values, to their nearest
dominating definition - either a reload, or the original definition.
Once this has been done, it then replaces all of the spill/reload
pseudo-ops with stores and loads, respectively. This has the effect of
splitting all of the live ranges we wanted to split, and determines how
the spills and reloads will be lowered.

NOTE: The spills analysis and transforms are not implemented in this
commit