feat(analysis): implement spills analysis
This analysis pass is used to determine what, if any, SSA values in a
given function must be spilled to memory, so as to ensure that the
operand stack depth at any given point in the function is within 16
elements. In addition to finding what values require spilling, it also
determines the placement of those spills, as well as corresponding
reloads, so that the values are on the operand stack when they are used.

The actual transformation is done separately, since attempting to do the
analysis and the transformation in a single pass introduces more
complexity, and makes it more difficult to identify where issues are
introduced. A separate analysis and transformation step ensures that we
can verify the analysis, and then verify that the transformation was
applied correctly.

NOTE: This commit does not contain the transformation passes that apply
the results of the spills analysis.